adapt CSG constructive operations tests to new project structure
